creating statsdash library folder and continuing with refactor
