+++ Statsdash/Templates/GA/base.html
@@ -1,0 +1,21 @@
+{% extends "../base.html" %}
+
+
+{% if warning_sites|length > 0 %}<p>We are missing some data for the following sites: {% for site in warning %}{{site|capita;ize}}, {% endfor %}</p>{% endif %}{% endif %}
+
+{# 
+{% if num_sites == total_num_sites %}{% if warning|length > 0 %}<p>We are missing some data for the following sites: {% for site in warning %}{{site|capfirst}}, {% endfor %}</p>{% endif %}{% endif %}
+{% if num_sites == 1 %}{% for site in warning %}{% if site in site_names %}<p>We are missing some data for this site.</p>{% endif %}{% endfor %}{% endif %}
+ #}    
+
+
+{% block body %}
+{% if summary_table %}{% include "summary_table.html" %}{% endif %}
+{% if geo_table %}{% include "geo_table.html" %}{% endif %}
+{% if top_articles %}{% include "article_table.html" %}{% endif %}
+{% if traffic %}{% include "traffic_table.html" %}{% endif %}
+{% if referrals %}{% include "referral_table.html" %}{% endif %}
+{% if device %}{% include "device_table.html" %}{% endif %}
+{% if social %}{% include "social_table.html" %}{% endif %}
+    
+{% endblock %}
