+++ Statsdash/Templates/Youtube/base.html
@@ -1,0 +1,149 @@
+{% extends "base.html" %}
+
+
+{% block body %}
+        {{ subject }}
+		<h2>Overall Performance</h2>
+		<p>Channels ranked by total watch time for this period</p>
+		<table border="0" cellspacing="0">
+			<thead>
+				<tr>
+					<th>Channel</th>
+					<th colspan="3">Watch Time</th>	
+					<th colspan="3">Subscriber Gain</th>	
+					<th colspan="3">Total Subscribers</th>
+				</tr>
+				<tr>
+					<th></th>
+					
+					<th>Figure</th>
+					<th>{{ change }}</th>
+					<th>YoY</th>
+					
+					<th>Figure</th>
+					<th>{{ change }}</th>
+					<th>YoY</th>
+					
+					<th>Figure</th>
+					<th>{{ change }}</th>					
+				</tr>
+			</thead>
+			<tbody>
+				{% for row in summary_table %}
+				<tr class="{{ loop.cycle('odd', 'even') }}">
+					<td>{{row.channel}}</td>
+					<td class="num">{{row.estimatedMinutesWatched|intcomma}}</td>
+					<td class="num" style="{% if row.MONTHLY_percentage_estimatedMinutesWatched > 0 %}color:green;{% else %}color:red;{% endif %}">{{row.MONTHLY_percentage_estimatedMinutesWatched|round}}%</td>
+					<td class="num" style="{% if row.YEARLY_percentage_estimatedMinutesWatched > 0 %}color:green;{% else %}color:red;{% endif %}">{{row.YEARLY_percentage_estimatedMinutesWatched|round}}%</td>
+										
+					<td class="num">{{row.subscriberChange|intcomma}}</td>
+					<td class="num" style="{% if row.MONTHLY_percentage_subscriberChange > 0 %}color:green;{% else %}color:red;{% endif %}">{{row.MONTHLY_percentage_subscriberChange|round}}%</td>
+					<td class="num" style="{% if row.YEARLY_percentage_subscriberChange > 0 %}color:green;{% else %}color:red;{% endif %}">{{row.YEARLY_percentage_subscriberChange|round}}%</td>
+					
+					<td class="num">{{row.subscriberCount|intcomma}}</td>
+					<td class="num" style="{% if row.MONTHLY_percentage_subscriberCount > 0 %}color:green;{% else %}color:red;{% endif %}">{{row.MONTHLY_percentage_subscriberCount|round}}%</td>
+				</tr>
+				{% endfor %}
+			</tbody>
+		</table>
+		
+		
+		<table border="0" cellspacing="0">
+			<thead>
+				<tr>
+					<th>Channel</th>
+					<th>Views</th>
+					<th>Likes:Dislikes Ratio</th>
+					<th>Likes/1000 views</th>
+					<th>Subs/1000 views</th>
+					<th>Comments/1000 views</th>
+					<th>Shares/1000 views</th>					
+				</tr>			
+			</thead>
+			<tbody>
+				{% for row in stats_table %}
+				<tr class="{{ loop.cycle('odd', 'even') }}">
+					<td>{{row.channel}}</td>
+					<td class="num">{{row.views|round|intcomma}}</td>
+					<td class="num">{{row.likeRatio|round|intcomma}}:{{row.dislikeRatio|round|intcomma}} (<span style="{% if row.MONTHLY_percentage_likeRatio > 0 %}color:green;{% else %}color:red;{% endif %}">{{row.MONTHLY_percentage_likeRatio|round|intcomma}}%</span>)</td>
+					<td class="num">{{row.likeRate|round|intcomma}} (<span style="{% if row.MONTHLY_percentage_likeRate > 0 %}color:green;{% else %}color:red;{% endif %}">{{row.MONTHLY_percentage_likeRate|round|intcomma}}%</span>)</td>																								
+					<td class="num">{{row.subsRate|round|intcomma}} (<span style="{% if row.MONTHLY_percentage_subsRate > 0 %}color:green;{% else %}color:red;{% endif %}">{{row.MONTHLY_percentage_subsRate|round|intcomma}}%</span>)</td>
+
+					<td class="num">{{row.commentRate|round|intcomma}} (<span style="{% if row.MONTHLY_percentage_commentRate > 0 %}color:green;{% else %}color:red;{% endif %}">{{row.MONTHLY_percentage_commentRate|round|intcomma}}%</span>)</td>
+
+					<td class="num">{{row.sharesRate|round|intcomma}} (<span style="{% if row.MONTHLY_percentage_sharesRate > 0 %}color:green;{% else %}color:red;{% endif %}">{{row.MONTHLY_percentage_sharesRate|round|intcomma}}%</span>)</td>					
+
+				</tr>
+				{% endfor %}
+			</tbody>
+		</table>
+		
+		<h2>Breakdown by Country</h2>
+		<table border="0" cellspacing="0">
+			<thead>
+				<tr>
+					<th>Country</th>
+					<th>Watch Time</th>
+					<th>Views</th>
+					<th>Subscriber Gain</th>
+				</tr>
+			</thead>
+			<tbody>
+				{% for row  in geo_table %}
+				<tr class="{{ loop.cycle('odd', 'even') }}">
+					<td>{{row.country}}</td>
+					<td class="num">{{row.estimatedMinutesWatched|round|intcomma}} (<span style="{% if row.MONTHLY_percentage_estimatedMinutesWatched > 0 %}color:green;{% else %}color:red;{% endif %}">{{row.MONTHLY_percentage_estimatedMinutesWatched|round|intcomma}}%</span>)</td>
+					<td class="num">{{row.views|round|intcomma}} (<span style="{% if row.MONTHLY_percentage_views > 0 %}color:green;{% else %}color:red;{% endif %}">{{row.MONTHLY_percentage_views|round|intcomma}}%</span>)</td>
+					<td class="num">{{row.subscriberChange|round|intcomma}} (<span style="{% if row.MONTHLY_percentage_subscriberChange > 0 %}color:green;{% else %}color:red;{% endif %}">{{row.MONTHLY_percentage_subscriberChange|round|intcomma}}%</span>)</td>
+				</tr>
+				{% endfor %}
+			</tbody>
+		</table>	
+		
+		<h2>Top 10 Videos</h2>
+		<p>Most popular videos this period, sorted by watch time.</p>
+		<table border="0" cellspacing="0">
+			<thead>
+				<tr>
+					<th>Channel</th>
+					<th>Video</th>
+					<th>Watch Time</th>
+					<th>Views</th>
+				</tr>
+			</thead>
+			<tbody>
+				{% for row  in top_vids %}
+				<tr class="{{ loop.cycle('odd', 'even') }}">
+					<td>{{row.channel}}</td>
+					<td><a href="https://www.youtube.com/watch?v={{row.video}}" target="_blank">{{row.title}}</a></td>
+					<td class="num">{{row.estimatedMinutesWatched|round|intcomma}}</td>
+					<td class="num">{{row.views|round|intcomma}}</td>
+				</tr>
+				{% endfor %}
+			</tbody>
+		</table>		
+
+		<h2>Traffic Sources</h2>
+		<p>Ranked by total watch time</p>
+		<table border="0" cellspacing="0">
+			<thead>
+				<tr>
+					<th>Channel</th>
+					{% for source in traffic.0 %}
+						<th>{{source.insightTrafficSourceType}}</th>
+					{%- endfor %}
+				</tr>
+			</thead>
+			<tbody>
+				{% for row in traffic %}
+				<tr class="{{ loop.cycle('odd', 'even') }}">
+					<td>{{row.0.channel}}</td>
+					{% for source in row %}
+                    <td class="num {{ source.insightTrafficSourceType }}">{{ source.source_percentage|round|intcomma }}%</td>
+                    {% endfor %}
+				</tr>
+				{% endfor %}
+			</tbody>
+		</table>	
+						
+{% endblock %}
