+++ Statsdash/Templates/site_dash.html
@@ -1,4 +1,3 @@
-{% load humanize %}
 <!DOCTYPE html>
 <html>
 	<head>
@@ -32,10 +31,6 @@
 	tr.even {
     	
 	}
-    span  {
-        display: list-item;
-        list-style-type: none;
-	}	
 	img {
     	margin: 0.2em;
 	}
