+++ Statsdash/Templates/template_sections/article_table.html
