+++ Statsdash/Templates/template_sections/country_summary_table.html
