+++ Statsdash/Templates/template_sections/overview_monthly_table.html
