+++ Statsdash/Templates/template_sections/site_summary_table.html
