+++ Statsdash/Templates/template_sections/social_network_table.html
