+++ Statsdash/Templates/template_sections/traffic_summary_table.html
