+++ Statsdash/Templates/widedash.html
