+++ Statsdash/Youtube/aggregate_data.py-oauth2.json
@@ -1,0 +1,1 @@
+{"_module": "oauth2client.client", "scopes": ["https://www.googleapis.com/auth/youtube.readonly", "https://www.googleapis.com/auth/yt-analytics.readonly", "https://www.googleapis.com/auth/youtubepartner"], "token_expiry": "2016-06-14T09:20:50Z", "id_token": null, "access_token": "ya29.CjAAA6QVeElGiCE1eJaj1FMRSbWKvtLWqwgeLvX57eRP2l8NxbqDbzPlV77zjE032bQ", "token_uri": "https://accounts.google.com/o/oauth2/token", "invalid": false, "token_response": {"access_token": "ya29.CjAAA6QVeElGiCE1eJaj1FMRSbWKvtLWqwgeLvX57eRP2l8NxbqDbzPlV77zjE032bQ", "token_type": "Bearer", "expires_in": 3600}, "client_id": "1378929557-laeaapvu1hfmr0aovonceg4hooma0c12.apps.googleusercontent.com", "token_info_uri": "https://www.googleapis.com/oauth2/v3/tokeninfo", "client_secret": "GgRwusQo7K0_0C0qPCGK29nW", "revoke_uri": "https://accounts.google.com/o/oauth2/revoke", "_class": "OAuth2Credentials", "refresh_token": "1/OlS4K57EFPRhjjugMQJVU45HJh9PA7Kg5iU3Dt8HVgk", "user_agent": null}
