+++ Statsdash/Youtube/client_secret_1378929557-laeaapvu1hfmr0aovonceg4hooma0c12.apps.googleusercontent.com.json
@@ -1,0 +1,1 @@
+{"installed":{"client_id":"1378929557-laeaapvu1hfmr0aovonceg4hooma0c12.apps.googleusercontent.com","project_id":"youtube-dash-1294","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://accounts.google.com/o/oauth2/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"GgRwusQo7K0_0C0qPCGK29nW","redirect_uris":["urn:ietf:wg:oauth:2.0:oob","http://localhost"]}}
