+++ Statsdash/Youtube/client_secrets.json
@@ -1,0 +1,11 @@
+{
+	"web":{
+		"client_id":"1378929557-ks110ik6q1d8802sv7egbjt9jaqsppok.apps.googleusercontent.com",
+		"project_id":"youtube-dash-1294",
+		"auth_uri":"https://accounts.google.com/o/oauth2/auth",
+		"token_uri":"https://accounts.google.com/o/oauth2/token",
+		"auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs",
+		"client_secret":"MwICg9GGSRCCIu85HxGRByWK",
+		"redirect_uris":["http://localhost:8080"]
+	}
+}
