+++ Statsdash/report_schedule.py-example
